+++ src/main/java/pl/carwebapp/model/Hatchback.java
@@ -1,9 +1,11 @@
 package pl.carwebapp.model;
 
+import java.math.BigDecimal;
+
 public class Hatchback extends AbstractCar {
 
     public Hatchback(String name, String type, int manufacturingYear, String plates, String vin, Transmission transmission, FuelType fuelType,
-                     TypeOfDrive typeOfDrive, int doors, int price, double mpg, int seats) {
+                     TypeOfDrive typeOfDrive, int doors, BigDecimal price, double mpg, int seats) {
         super(type, name, manufacturingYear, Segment.C, plates, vin, transmission, fuelType, typeOfDrive, doors, price, mpg, seats);
     }
 }
