+++ src/main/java/pl/carwebapp/model/Prestige.java
@@ -1,9 +1,11 @@
 package pl.carwebapp.model;
 
+import java.math.BigDecimal;
+
 public class Prestige extends AbstractCar {
 
     public Prestige(String name, String type, int manufacturingYear, String plates, String vin, Transmission transmission, FuelType fuelType,
-                    TypeOfDrive typeOfDrive, int doors, int price, double mpg, int seats) {
+                    TypeOfDrive typeOfDrive, int doors, BigDecimal price, double mpg, int seats) {
         super(type, name, manufacturingYear, Segment.F, plates, vin, transmission, fuelType, typeOfDrive, doors, price, mpg, seats);
     }
 }
